Refactor `fake_commit` into test.helpers.
